% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_tile_corners.R
\name{calc_tile_corners}
\alias{calc_tile_corners}
\title{Calculate the tile corner XY values in source raster/image grid}
\usage{
calc_tile_corners(source_n_rows, source_n_cols, tile_n_rows, tile_n_cols,
  tile_overlap = 0, complete_image = FALSE)
}
\arguments{
\item{source_n_rows}{Number of rows in source raster/image}

\item{source_n_cols}{Number of columns in source raster/image}

\item{tile_n_rows}{Number of rows in each tile}

\item{tile_n_cols}{Number of columns in each tile}

\item{tile_overlap}{Number of pixel overlap in adjacent tiles (in both X and Y directions)}

\item{complete_image}{If TRUE and the tile size and overlap dimensions do not conform to
covering the entire source raster/image, an additional row and column
of tiles will be created that include the excluded pixels but do NOT
respect the overlap value. If FALSE and the dimensions do not conform,
the set of tiles will omit some pixels on the right and bottom side
of the source raster/image}
}
\value{
Data frame with corner cells of one tile in each row, containing
variables \['x0', 'x1', 'y0', 'y1'\] for start \(\*0\\) and stop \(\*1)
cell numbers in horizontal \(x\*\) and vertical \(y\*\) directions
}
\description{
Calculate the tile corner XY values in source raster/image grid
}
