% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_tile_data.R
\name{score_tile_data}
\alias{score_tile_data}
\title{Score a sliced tile data frame with a Keras model}
\usage{
score_tile_data(tile_data, model_params, scoring_model,
  score_outpath = NULL, compress_score_rds = FALSE,
  return_score = TRUE, keep_array = FALSE, verbose = FALSE)
}
\arguments{
\item{tile_data}{Data frame containing image slice data}

\item{model_params}{Model parameters object created during model training}

\item{scoring_model}{Trained keras model object}

\item{score_outpath}{Path to save results of image scoring as .rds file}

\item{compress_score_rds}{Should saved score results be compressed}

\item{return_score}{Should score results be returned as an R object by the function}

\item{verbose}{Should messages about current step being processes be printed to screen?}
}
\value{
List containing model parameters and input `img_data` data frame with
additional column containing model predicted probabilities for each slice

If score \code{outpath != NULL} returned data frame will be saved to outpath as an
RDS file, compressed if \code{compress_score_rds == TRUE}.
}
\description{
Score a sliced tile data frame with a Keras model
}
