% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_image.R
\name{slice_image}
\alias{slice_image}
\title{Slice an image into tiles for scoring with a Keras model}
\usage{
slice_image(img_path, slice_n_rows, slice_n_cols, slice_overlap_px = 0,
  complete_image = FALSE, img_xml_wkt_path = NULL, wkt_string = NULL,
  proj4_string = NULL, project_raster = FALSE, verbose = FALSE)
}
\arguments{
\item{img_path}{Path to image file that can be read by \code{\link[raster]{brick}}.}

\item{slice_n_rows}{Number of rows in each tile. See \code{\link{calc_tile_corners}}.}

\item{slice_n_cols}{Number of columns in each tile. See \code{\link{calc_tile_corners}}.}

\item{slice_overlap_px}{Number of pixel overlap in adjacent tiles (in both X and Y directions).
See \code{\link{calc_tile_corners}}.}

\item{complete_image}{If TRUE and the tile size and overlap dimensions do not conform to
covering the entire source raster/image, an additional row and column
of tiles will be created that include the excluded pixels but do NOT
respect the overlap value. If FALSE and the dimensions do not conform,
the set of tiles will omit some pixels on the right and bottom side
of the source raster/image. See \code{\link{calc_tile_corners}}.}

\item{img_xml_wkt_path}{Optional path to .xml file containing metadata about
image in WKT format.}

\item{wkt_string}{Optional WKT string containing image projection information.
Overridden by `img_xml_wkt_path` if present.}

\item{proj4_string}{Optional proj4 string containing image projection information.
Overriden by `img_xml_wkt_path` of `wkt_string` if present.}

\item{verbose}{Should messages about current step being processes be printed to screen?}
}
\value{
Data frame with one row for each tile, containing:
- \code{img_path}
- \code{img_xml_wkt_path} (if present)
- Numeric tile ID number
- Tile corner cells in source jp2 pixel values
- List-column containing each tile's extent based on the source jp2's projected raster
- sf geometry column containing each tile's sf polygon
- List-column containing tile's RGB layers in a 4d array of dimension
[1, \code{slice_n_cols}, \code{slice_n_rows}, 3]
}
\description{
Slice an image into tiles for scoring with a Keras model
}
